refactor: Split into multiple files and handle global state better
