+++ lua/dap/constants.lua
@@ -1,0 +1,4 @@
+
+return {
+  ns_pos = 'dap_pos'
+}
