+++ lua/dap/rpc.lua
@@ -1,0 +1,97 @@
+
+-- Copied from neovim rpc.lua
+local function parse_headers(header)
+  if type(header) ~= 'string' then
+    return nil
+  end
+  local headers = {}
+  for line in vim.gsplit(header, '\r\n', true) do
+    if line == '' then
+      break
+    end
+    local key, value = line:match('^%s*(%S+)%s*:%s*(.+)%s*$')
+    if key then
+      key = key:lower():gsub('%-', '_')
+      headers[key] = value
+    else
+      error(string.format("Invalid header line %q", line))
+    end
+  end
+  headers.content_length = tonumber(headers.content_length)
+    or error(string.format("Content-Length not found in headers. %q", header))
+  return headers
+end
+
+-- Mostly copied from neovim rpc.lua
+local header_start_pattern = ("content"):gsub("%w", function(c) return "["..c..c:upper().."]" end)
+local function parse_chunk_loop()
+  local buffer = ''
+  while true do
+    local start, finish = buffer:find('\r\n\r\n', 1, true)
+    if start then
+      local buffer_start = buffer:find(header_start_pattern)
+      local headers = parse_headers(buffer:sub(buffer_start, start - 1))
+      local content_length = headers.content_length
+      local body_chunks = {buffer:sub(finish + 1)}
+      local body_length = #body_chunks[1]
+      while body_length < content_length do
+        local chunk = coroutine.yield()
+          or error("Expected more data for the body. The server may have died.")
+        table.insert(body_chunks, chunk)
+        body_length = body_length + #chunk
+      end
+      local last_chunk = body_chunks[#body_chunks]
+
+      body_chunks[#body_chunks] = last_chunk:sub(1, content_length - body_length - 1)
+      local rest = ''
+      if body_length > content_length then
+        rest = last_chunk:sub(content_length - body_length)
+      end
+      local body = table.concat(body_chunks)
+      buffer = rest .. (coroutine.yield(headers, body)
+        or error("Expected more data for the body. The server may have died."))
+    else
+      buffer = buffer .. (coroutine.yield()
+        or error("Expected more data for the header. The server may have died."))
+    end
+  end
+end
+
+local function create_read_loop(handle_body)
+  local parse_chunk = coroutine.wrap(parse_chunk_loop)
+  parse_chunk()
+  return function (err, chunk)
+    if err then
+      print(err)
+      return
+    end
+    if not chunk then
+      return
+    end
+    while true do
+      local headers, body = parse_chunk(chunk)
+      if headers then
+        vim.schedule(function()
+          handle_body(body)
+        end)
+        chunk = ''
+      else
+        break
+      end
+    end
+  end
+end
+
+local function msg_with_content_length(msg)
+  return table.concat {
+    'Content-Length: ';
+    tostring(#msg);
+    '\r\n\r\n';
+    msg
+  }
+end
+
+return {
+  create_read_loop = create_read_loop,
+  msg_with_content_length = msg_with_content_length,
+}
